WIP: started to implement static color change
